Discord/requirements.txt
discord.py
aiohttp
python-dotenv
python-dateutil
pytz
json5
dotenv
